cleaning it up

+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/exchange/model/DexOrderDBRequestForTrading.java
@@ -18,8 +18,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DexOrderDBRequestForTrading {
-    private long startInterval;
-    private long endInterval;
+    private Integer startInterval;
+    private Integer endInterval;
     private byte pairCur;
     private Integer offset;
     private Integer limit;
